+++ lib/new_relic/agent/transaction/request_attributes.rb
@@ -12,8 +12,6 @@ module NewRelic
                     :host, :port, :user_agent, :request_method
 
         HTTP_ACCEPT_HEADER_KEY = "HTTP_ACCEPT".freeze
-        REQUEST_URI_KEY = "request_uri".freeze
-        WILDCARD = "*".freeze
 
         def initialize request
           @request_path = path_from_request request
@@ -36,12 +34,11 @@ module NewRelic
             txn.add_agent_attribute :'request.headers.referer', referer, AttributeFilter::DST_ERROR_COLLECTOR
           end
 
-          # This is temporary and aims to avoid collecting this attribute on transaction and error traces, 
-          # which already directly have a path value that the RPM UI depends on. We will either only
-          # collect request_uri as an agent attribute, in conjunction with UI work, or we will stop collecting
-          # this as an agent attribute (RUBY-1573)
-          if request_path && configured_to_collect?
-            txn.add_agent_attribute :request_uri, request_path, AttributeFilter::DST_TRANSACTION_EVENTS
+          if request_path
+            txn.add_agent_attribute :'request.uri',
+                                    request_path,
+                                    AttributeFilter::DST_TRANSACTION_TRACER |
+                                    AttributeFilter::DST_ERROR_COLLECTOR
           end
 
           if accept
@@ -118,11 +115,6 @@ module NewRelic
           if env = attribute_from_request(request, :env)
             env[key]
           end
-        end
-
-        def configured_to_collect?
-          txn_event_attributes = NewRelic::Agent.config[:'transaction_events.attributes.include']
-          txn_event_attributes.any?{|attribute| attribute == REQUEST_URI_KEY || attribute == WILDCARD}
         end
       end
     end