+++ test/multiverse/suites/rails/parameter_capture_test.rb
@@ -49,14 +49,15 @@ class ParameterCaptureTest < ActionDispatch::IntegrationTest
   def test_uri_on_traced_errors_never_contains_query_string_without_capture_params
     with_config(:capture_params => false) do
       get '/parameter_capture/error?other=1234&secret=4567'
-      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("request_uri"))
+
+      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("agentAttributes")["request.uri"])
     end
   end
 
   def test_uri_on_traced_errors_never_contains_query_string_with_capture_params
     with_config(:capture_params => true) do
       get '/parameter_capture/error?other=1234&secret=4567'
-      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("request_uri"))
+      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("agentAttributes")["request.uri"])
     end
   end
 
@@ -102,12 +103,16 @@ class ParameterCaptureTest < ActionDispatch::IntegrationTest
     with_config(:capture_params => false) do
       get '/parameter_capture/transaction?other=1234&secret=4567'
     end
-    assert_equal('/parameter_capture/transaction', last_transaction_trace.uri)
+
+    attributes = last_transaction_trace.attributes.agent_attributes_for(NewRelic::Agent::AttributeFilter::DST_TRANSACTION_TRACER)
+    assert_equal('/parameter_capture/transaction', attributes[:'request.uri'])
 
     with_config(:capture_params => true) do
       get '/parameter_capture/transaction?other=1234&secret=4567'
     end
-    assert_equal('/parameter_capture/transaction', last_transaction_trace.uri)
+
+    attributes = last_transaction_trace.attributes.agent_attributes_for(NewRelic::Agent::AttributeFilter::DST_TRANSACTION_TRACER)
+    assert_equal('/parameter_capture/transaction', attributes[:'request.uri'])
   end
 
   def test_filters_parameters_on_traced_errors
@@ -163,27 +168,29 @@ class ParameterCaptureTest < ActionDispatch::IntegrationTest
     with_config(:capture_params => false) do
       get '/parameter_capture/transaction?param1=value1&param2=value2'
     end
-    assert_equal('/parameter_capture/transaction', last_transaction_trace.uri)
+    attributes = last_transaction_trace.attributes.agent_attributes_for(NewRelic::Agent::AttributeFilter::DST_TRANSACTION_TRACER)
+    assert_equal('/parameter_capture/transaction', attributes[:'request.uri'])
   end
 
   def test_uri_on_tt_should_not_contain_query_string_with_capture_params_on
     with_config(:capture_params => true) do
       get '/parameter_capture/transaction?param1=value1&param2=value2'
     end
-    assert_equal('/parameter_capture/transaction', last_transaction_trace.uri)
+    attributes = last_transaction_trace.attributes.agent_attributes_for(NewRelic::Agent::AttributeFilter::DST_TRANSACTION_TRACER)
+    assert_equal('/parameter_capture/transaction', attributes[:'request.uri'])
   end
 
   def test_uri_on_traced_error_should_not_contain_query_string_with_capture_params_off
     with_config(:capture_params => false) do
       get '/parameter_capture/error?param1=value1&param2=value2'
-      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("request_uri"))
+      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("agentAttributes")["request.uri"])
     end
   end
 
   def test_uri_on_traced_error_should_not_contain_query_string_with_capture_params_on
     with_config(:capture_params => true) do
       get '/parameter_capture/error?param1=value1&param2=value2'
-      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("request_uri"))
+      assert_equal('/parameter_capture/error', attributes_for_single_error_posted("agentAttributes")["request.uri"])
     end
   end